Flytter periodeElement ut i egen fil.
