+++ src/sider/ftrl/saksbehandling/initialStegs.tsx
@@ -2,7 +2,7 @@ import { FANE_STATUS } from "../../../felleskomponenter/stegvelger";
 import { VurderingStart } from "./stegKomponenter/vurderingStart";
 import { VurderingVirksomhet } from "./stegKomponenter/vurderingVirksomhet";
 import { VurderingBestemmelse } from "./stegKomponenter/vurderingBestemmelse";
-import { VurderingPerioder } from "./stegKomponenter/vurderingPerioder";
+import { VurderingPerioder } from "./stegKomponenter/vurderingPeriode/vurderingPerioder";
 import { VurderingTrygdeavgift } from "./stegKomponenter/vurderingTrygdeavgift";
 import { VurderingVedtak } from "./stegKomponenter/vurderingVedtak";
