+++ src/sider/ftrl/saksbehandling/stegKomponenter/vurderingPeriode/vurderingPerioder.less
